Move some supplementary functions to utility.py
